requirements.txt
opencv-python
numpy
av
pyyaml
ruamel.yaml
pillow
loguru
pyautogui
python-xlib
loguru
tqdm
scipy
halo
git+https://github.com/wyatthuckaby/python-uinput.git
git+https://github.com/lilohuang/PyTurboJPEG.git
python-uinput
psycopg[binary,pool]
PyWinCtl
git+https://github.com/Adelaide-Autonomous-Racing-Challenge/ac-state.git
